Rename `rules` package to `rule`
