+++ base/src/main/java/io/spine/validate/rule/ValidationRules.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.validate.rules;
+package io.spine.validate.rule;
 
 import com.google.common.base.Splitter;
 import com.google.common.collect.ImmutableCollection;
